refactor add exclude_filter.py
